Add Load Manager setup
This setup allows for more control over game elements as they happen. The flow of the game is as follows:
1. The starter script's OnSceneLoad will be called when the game is started. This will create a Game Manager object that contains the Load Manager script and starts the StartRoutine from the LoadManager.
2. The StartRoutine in the LoadManager will instantiate the player and camera and load the loading scene.
3. Depending on the current scene, a different coroutine will start.

If the current scene is the StartScene, then we start by enabling and disabling the loading screen. We will also deactivate the player input here so the player isn't moving in the background. Once this scene is loaded, the logo has an animation that calls GoToMenu after the logo appears.
The GoToMenu routine will enable the loading screen, wait a little bit so its not too sudden then, unloads open scenes and loads the MainMenu Scene.
If the Start Game Button is pressed, the MainMenu Scene is unloaded and the Game Level is opened.