Added Level 1 of the game
New edited game assets were added to make the appropriate changes to the prototype level.
While the prototype level was broken down into smaller level to fit the length of the first level.
New Tiles were added to the tile palette for the team to use along with new prefabs for the first level along with previous versions of it.
Finally the collision of player with the wall was fixed to ensure smooth gameplay.