Add Ability to Switch Skins
There are three skins the player can have; Fire, Frost, and Lightning. The player can switch between these skins using the 1, 2, 3 keys. This happens via the Elements, Skins, Arms, and Bullets array where in each array in the same order, there is the Sprite/Prefab that represents that element. For example. the first element in all arrays is Fire.

The player can also switch between two weapons. I have added an Abstract weapon class since the only difference between weapons is the way they Fire.

The bullets act differently based on the element. This is yet to be implemented.