Added a functionality to level select screen
Adjusted the Load Manager Script to allow the level 2, and level 3 to be loaded in the coroutine as well when called. While the Main Menu Manger was also updated to allow to call the Go to Level functions for each level which was then attached to each buttons. A back button is also added to easily navigate between main menu and level select screen.