Audio Implementation
I've added an AudioManager script responsible for managing audio in the game. This manager is equipped with methods to play various sounds at specific events:

PlayPlayerShootSound(): Plays a sound when the player shoots.

PlayPlayerHitSound(): Plays a sound when the player is hit.

PlayPlayerDeathSound(): Plays a sound when the player dies.

PlayEnemyDefeatSound(): Plays a sound when an enemy is defeated.

PlayPlayerSwitchWeaponSound(): Plays a sound when the player switches weapons.