Add Fire and Frost Elements and Enemy Health Bar
This commit adds fire and lightning powers. When the element bar gets to its peak, the fire element deal continuous damage to the enemy and its surrounding allies, while the lightning power deals instant damage to the enemy and allies.

To better visualize enemy damage, we have also added an enemy health bar.